created a class createPatietView
A class "createPatientView" was created in which new patients can be created and then saved with the "save" button or cancelled with "cancel"

A button "create Patient" is created in the UserView class, which opens a window in which the patient's data can be entered.

In the UserController class, an ActionListener is programmed for the newly created "save" and "cancel" buttons.